refactor: step3의 리뷰사항 반영 커밋
- LottoResultPack: 로또 매칭 결과의 개수에 따른 결과 (LottoResult) 를 미리 Option 클래스에서 만들어둔 해시맵에 매칭된 것을 가져오는데,
  이 때, 키를 사용한 순회자 iterator 대신, values() 를 사용해서 가독성을 높임.
  구조는 동일함, 개수별 결과가 미리 맵으로 있고, 게임하면서 로또 매칭의 결과는 각 결과별로 몇장의 로또들이 해당 되어있는지 카운팅 한 합이 있다.
- UtilsString -> StringUtils 클래스명 변경 (Utils를 뒤로 붙이는게 일반적)
