+++ src/main/java/calculator/domain/Numbers.java
@@ -3,8 +3,6 @@ package calculator.domain;
 import java.util.ArrayList;
 import java.util.List;
 
-import static calculator.utils.UtilsString.isNullOrEmpty;
-
 public class Numbers {
     List<Number> numbers = new ArrayList<>();
