+++ src/main/java/calculator/utils/StringUtils.java
@@ -1,6 +1,6 @@
 package calculator.utils;
 
-public class UtilsString {
+public class StringUtils {
 
     public static boolean isNullOrEmpty(String input) {
         return input == null || input.isEmpty();
